Mirror ticks #30
This feature here basically looks at all the climbs a user has done and sees if they completed it. If they complete the climb, it gets checked off. If they complete the climb in reverse, then then a reversed checkmark shows. However, if a user marks the climbs in the forward and reverse way, then a forward and reflected check mark shows.